+++ cli/azd/cmd/monitor.go
@@ -6,6 +6,8 @@ package cmd
 import (
 	"context"
 	"fmt"
+	"os"
+	"os/exec"
 
 	"github.com/azure/azure-dev/cli/azd/internal"
 	"github.com/azure/azure-dev/cli/azd/pkg/environment/azdcontext"
@@ -13,7 +15,7 @@ import (
 	"github.com/azure/azure-dev/cli/azd/pkg/input"
 	"github.com/azure/azure-dev/cli/azd/pkg/tools"
 	"github.com/azure/azure-dev/cli/azd/pkg/tools/azcli"
-	"github.com/pbnj/go-open"
+	"github.com/cli/browser"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 )
@@ -140,7 +142,21 @@ func (m *monitorAction) Run(ctx context.Context) error {
 	openWithDefaultBrowser := func(url string) {
 		fmt.Fprintf(m.console.Handles().Stdout, "Opening %s in the default browser...\n", url)
 
-		if err := open.Open(url); err != nil {
+		// In Codespaces and devcontainers a $BROWSER environment variable is
+		// present whose value is an executable that launches the browser when
+		// called with the form:
+		// $BROWSER <url>
+
+		const BrowserEnvVarName = "BROWSER"
+
+		if envBrowser := os.Getenv(BrowserEnvVarName); len(envBrowser) > 0 {
+			if err := exec.Command(envBrowser, url).Run(); err != nil {
+				fmt.Fprintf(m.console.Handles().Stderr, "warning: failed to open browser configured by $BROWSER: %s\n", err.Error())
+			}
+			return
+		}
+
+		if err := browser.OpenURL(url); err != nil {
 			fmt.Fprintf(m.console.Handles().Stderr, "warning: failed to open default browser: %s\n", err.Error())
 		}
 	}