+++ cli/azd/pkg/infra/provisioning/bicep/bicep_provider_test.go
@@ -9,18 +9,23 @@ import (
 	"encoding/json"
 	"io"
 	"net/http"
+	"os"
+	"path"
 	"strings"
 	"testing"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/appconfiguration/armappconfiguration"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/keyvault/armkeyvault"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/resources/armresources"
+	"github.com/azure/azure-dev/cli/azd/pkg/azure"
 	"github.com/azure/azure-dev/cli/azd/pkg/convert"
 	"github.com/azure/azure-dev/cli/azd/pkg/environment"
 	"github.com/azure/azure-dev/cli/azd/pkg/exec"
 	"github.com/azure/azure-dev/cli/azd/pkg/infra"
 	. "github.com/azure/azure-dev/cli/azd/pkg/infra/provisioning"
 	"github.com/azure/azure-dev/cli/azd/pkg/input"
-	"github.com/azure/azure-dev/cli/azd/pkg/tools/azcli"
+	"github.com/azure/azure-dev/cli/azd/pkg/osutil"
 	"github.com/azure/azure-dev/cli/azd/test/mocks"
 	execmock "github.com/azure/azure-dev/cli/azd/test/mocks/exec"
 	"github.com/azure/azure-dev/cli/azd/test/mocks/httputil"
@@ -34,7 +39,7 @@ func TestBicepPlan(t *testing.T) {
 
 	mockContext := mocks.NewMockContext(context.Background())
 	prepareGenericMocks(mockContext.CommandRunner)
-	preparePlanningMocks(mockContext.CommandRunner)
+	preparePlanningMocks(mockContext)
 	prepareDeployShowMocks(mockContext.HttpClient)
 	infraProvider := createBicepProvider(*mockContext.Context)
 	planningTask := infraProvider.Plan(*mockContext.Context)
@@ -78,7 +83,7 @@ func TestBicepState(t *testing.T) {
 
 	mockContext := mocks.NewMockContext(context.Background())
 	prepareGenericMocks(mockContext.CommandRunner)
-	preparePlanningMocks(mockContext.CommandRunner)
+	preparePlanningMocks(mockContext)
 	prepareDeployShowMocks(mockContext.HttpClient)
 	prepareDeployMocks(mockContext.CommandRunner)
 
@@ -125,14 +130,20 @@ func TestBicepDeploy(t *testing.T) {
 
 	mockContext := mocks.NewMockContext(context.Background())
 	prepareGenericMocks(mockContext.CommandRunner)
-	preparePlanningMocks(mockContext.CommandRunner)
+	preparePlanningMocks(mockContext)
 	prepareDeployShowMocks(mockContext.HttpClient)
 	prepareDeployMocks(mockContext.CommandRunner)
 
 	infraProvider := createBicepProvider(*mockContext.Context)
+	tmpPath := t.TempDir()
+	parametersPath := path.Join(tmpPath, "params.json")
+	createTmpFile := os.WriteFile(parametersPath, []byte(testArmParametersFile), osutil.PermissionFile)
+	require.NoError(t, createTmpFile)
+
 	deploymentPlan := DeploymentPlan{
 		Details: BicepDeploymentDetails{
-			ParameterFilePath: "",
+			ParameterFilePath: parametersPath,
+			Template:          to.Ptr(azure.ArmTemplate("{}")),
 		},
 	}
 
@@ -169,7 +180,7 @@ func TestBicepDestroy(t *testing.T) {
 	t.Run("Interactive", func(t *testing.T) {
 		mockContext := mocks.NewMockContext(context.Background())
 		prepareGenericMocks(mockContext.CommandRunner)
-		preparePlanningMocks(mockContext.CommandRunner)
+		preparePlanningMocks(mockContext)
 		prepareDeployShowMocks(mockContext.HttpClient)
 		prepareDestroyMocks(mockContext)
 
@@ -183,7 +194,7 @@ func TestBicepDestroy(t *testing.T) {
 		}).Respond(true)
 
 		mockContext.Console.WhenConfirm(func(options input.ConsoleOptions) bool {
-			return strings.Contains(options.Message, "Would you like to permanently delete these Key Vaults")
+			return strings.Contains(options.Message, "Would you like to permanently delete these Key Vaults/App Configurations")
 		}).Respond(true)
 
 		infraProvider := createBicepProvider(*mockContext.Context)
@@ -213,28 +224,31 @@ func TestBicepDestroy(t *testing.T) {
 
 		// Verify console prompts
 		consoleOutput := mockContext.Console.Output()
-		require.Len(t, consoleOutput, 6)
+		require.Len(t, consoleOutput, 7)
 		require.Contains(t, consoleOutput[0], "This will delete")
 		require.Contains(t, consoleOutput[1], "Deleted resource group")
-		require.Contains(t, consoleOutput[2], "This operation will delete and purge")
-		require.Contains(t, consoleOutput[3], "Would you like to permanently delete these Key Vaults")
+		require.Contains(t, consoleOutput[2], "This operation will delete")
+		require.Contains(t, consoleOutput[3], "Would you like to permanently delete these Key Vaults/App Configurations")
 		require.Contains(t, consoleOutput[4], "Purged key vault")
-		require.Contains(t, consoleOutput[5], "Deleted deployment")
+		require.Contains(t, consoleOutput[5], "Purged app configuration")
+		require.Contains(t, consoleOutput[6], "Deleted deployment")
 
 		// Verify progress output
-		require.Len(t, progressLog, 6)
+		require.Len(t, progressLog, 8)
 		require.Contains(t, progressLog[0], "Fetching resource groups")
 		require.Contains(t, progressLog[1], "Fetching resources")
-		require.Contains(t, progressLog[2], "Getting KeyVaults to purge")
-		require.Contains(t, progressLog[3], "Deleting resource group")
-		require.Contains(t, progressLog[4], "Purging key vault")
-		require.Contains(t, progressLog[5], "Deleting deployment")
+		require.Contains(t, progressLog[2], "Getting Key Vaults to purge")
+		require.Contains(t, progressLog[3], "Getting App Configurations to purge")
+		require.Contains(t, progressLog[4], "Deleting resource group")
+		require.Contains(t, progressLog[5], "Purging key vault")
+		require.Contains(t, progressLog[6], "Purging app configuration")
+		require.Contains(t, progressLog[7], "Deleting deployment")
 	})
 
 	t.Run("InteractiveForceAndPurge", func(t *testing.T) {
 		mockContext := mocks.NewMockContext(context.Background())
 		prepareGenericMocks(mockContext.CommandRunner)
-		preparePlanningMocks(mockContext.CommandRunner)
+		preparePlanningMocks(mockContext)
 		prepareDeployShowMocks(mockContext.HttpClient)
 		prepareDestroyMocks(mockContext)
 
@@ -269,18 +283,17 @@ func TestBicepDestroy(t *testing.T) {
 
 		// Verify console prompts
 		consoleOutput := mockContext.Console.Output()
-		require.Len(t, consoleOutput, 3)
+		require.Len(t, consoleOutput, 2)
 		require.Contains(t, consoleOutput[0], "Deleted resource group")
-		require.Contains(t, consoleOutput[1], "Purged key vault")
-		require.Contains(t, consoleOutput[2], "Deleted deployment")
+		require.Contains(t, consoleOutput[1], "Deleted deployment")
 
 		// Verify progress output
 		require.Len(t, progressLog, 6)
 		require.Contains(t, progressLog[0], "Fetching resource groups")
 		require.Contains(t, progressLog[1], "Fetching resources")
-		require.Contains(t, progressLog[2], "Getting KeyVaults to purge")
-		require.Contains(t, progressLog[3], "Deleting resource group")
-		require.Contains(t, progressLog[4], "Purging key vault")
+		require.Contains(t, progressLog[2], "Getting Key Vaults to purge")
+		require.Contains(t, progressLog[3], "Getting App Configurations to purge")
+		require.Contains(t, progressLog[4], "Deleting resource group")
 		require.Contains(t, progressLog[5], "Deleting deployment")
 	})
 }
@@ -330,8 +343,7 @@ func prepareDeployMocks(commandRunner *execmock.MockCommandRunner) {
 }
 
 func preparePlanningMocks(
-	commandRunner *execmock.MockCommandRunner) {
-	expectedWebsiteUrl := "http://myapp.azurewebsites.net"
+	mockContext *mocks.MockContext) {
 	bicepInputParams := make(map[string]BicepInputParameter)
 	bicepInputParams["environmentName"] = BicepInputParameter{Value: "${AZURE_ENV_NAME}"}
 	bicepInputParams["location"] = BicepInputParameter{Value: "${AZURE_LOCATION}"}
@@ -343,43 +355,37 @@ func preparePlanningMocks(
 		Outputs:    bicepOutputParams,
 	}
 
-	deployOutputs := make(map[string]azcli.AzCliDeploymentOutput)
-	deployOutputs["WEBSITE_URL"] = azcli.AzCliDeploymentOutput{Type: "String", Value: expectedWebsiteUrl}
-	azDeployment := azcli.AzCliDeployment{
-		Id:   "DEPLOYMENT_ID",
-		Name: "DEPLOYMENT_NAME",
-		Properties: azcli.AzCliDeploymentProperties{
-			Outputs: deployOutputs,
-			Dependencies: []azcli.AzCliDeploymentPropertiesDependency{
-				{
-					DependsOn: []azcli.AzCliDeploymentPropertiesBasicDependency{
-						{
-							Id:           "RESOURCE_ID",
-							ResourceName: "RESOURCE_GROUP",
-							ResourceType: string(infra.AzureResourceTypeResourceGroup),
-						},
-					},
-				},
-			},
-		},
-	}
-
 	bicepBytes, _ := json.Marshal(bicepTemplate)
-	deployResultBytes, _ := json.Marshal(azDeployment)
+	deployResult := `
+	{
+		"id":"DEPLOYMENT_ID",
+		"name":"DEPLOYMENT_NAME",
+		"properties":{
+			"outputs":{
+				"WEBSITE_URL":{"type": "String", "value": "http://myapp.azurewebsites.net"}
+			}
+		}
+	}`
 
-	commandRunner.When(func(args exec.RunArgs, command string) bool {
+	mockContext.CommandRunner.When(func(args exec.RunArgs, command string) bool {
 		return strings.Contains(command, "az bicep build")
 	}).Respond(exec.RunResult{
 		Stdout: string(bicepBytes),
 		Stderr: "",
 	})
-
-	// ARM deployment
-	commandRunner.When(func(args exec.RunArgs, command string) bool {
-		return strings.Contains(command, "az deployment sub create")
-	}).Respond(exec.RunResult{
-		Stdout: string(deployResultBytes),
-		Stderr: "",
+	mockContext.HttpClient.When(func(request *http.Request) bool {
+		return request.Method == http.MethodPut && strings.Contains(
+			request.URL.Path,
+			"/subscriptions/SUBSCRIPTION_ID/providers/Microsoft.Resources/deployments",
+		)
+	}).RespondFn(func(request *http.Request) (*http.Response, error) {
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       io.NopCloser(bytes.NewBuffer([]byte(deployResult))),
+			Request: &http.Request{
+				Method: http.MethodGet,
+			},
+		}, nil
 	})
 }
 
@@ -439,6 +445,12 @@ func prepareDestroyMocks(mockContext *mocks.MockContext) {
 				Type:     convert.RefOf(string(infra.AzureResourceTypeKeyVault)),
 				Location: convert.RefOf("eastus2"),
 			},
+			{
+				ID:       convert.RefOf("appconfiguration"),
+				Name:     convert.RefOf("ac-123"),
+				Type:     convert.RefOf(string(infra.AzureResourceTypeAppConfig)),
+				Location: convert.RefOf("eastus2"),
+			},
 		},
 	}
 
@@ -478,6 +490,29 @@ func prepareDestroyMocks(mockContext *mocks.MockContext) {
 		}, nil
 	})
 
+	// Get App Configuration
+	mockContext.HttpClient.When(func(request *http.Request) bool {
+		return request.Method == http.MethodGet && strings.Contains(request.URL.Path, "/configurationStores/ac-123")
+	}).RespondFn(func(request *http.Request) (*http.Response, error) {
+		appConfigResponse := armappconfiguration.ConfigurationStoresClientGetResponse{
+			ConfigurationStore: armappconfiguration.ConfigurationStore{
+				ID:       convert.RefOf("ac-123"),
+				Name:     convert.RefOf("ac-123"),
+				Location: convert.RefOf("eastus2"),
+				Properties: &armappconfiguration.ConfigurationStoreProperties{
+					EnablePurgeProtection: convert.RefOf(false),
+				},
+			},
+		}
+
+		appConfigBytes, _ := json.Marshal(appConfigResponse)
+
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       io.NopCloser(bytes.NewBuffer(appConfigBytes)),
+		}, nil
+	})
+
 	// Delete resource group
 	mockContext.HttpClient.When(func(request *http.Request) bool {
 		return request.Method == http.MethodDelete &&
@@ -503,6 +538,19 @@ func prepareDestroyMocks(mockContext *mocks.MockContext) {
 		}, nil
 	})
 
+	// Purge App configuration
+	mockContext.HttpClient.When(func(request *http.Request) bool {
+		return request.Method == http.MethodPost && strings.Contains(request.URL.Path,
+			"deletedConfigurationStores/ac-123/purge")
+	}).RespondFn(func(request *http.Request) (*http.Response, error) {
+		return &http.Response{
+			Request:    request,
+			Header:     http.Header{},
+			StatusCode: http.StatusOK,
+			Body:       http.NoBody,
+		}, nil
+	})
+
 	// Delete deployment
 	mockContext.CommandRunner.When(func(args exec.RunArgs, command string) bool {
 		return strings.Contains(command, "az deployment sub delete")
@@ -511,3 +559,11 @@ func prepareDestroyMocks(mockContext *mocks.MockContext) {
 		Stderr: "",
 	})
 }
+
+var testArmParametersFile string = `{
+	"parameters": {
+		"location": {
+			"value": "West US"
+		}
+	}
+}`